requirements_whisper.txt
google-api-python-client>=2.118.0
youtube-transcript-api>=1.2.3
pytube>=15.0.0
openai-whisper>=20250625
sumy>=0.11.0
nltk>=3.8.1
